Split h261 frames into rtp packets in codec
Previously the encoder was told to produce very (64k) large rtp packets
(to ensure each frame was contained in one packet)
so that the framework could split and merge the encoded frames properly.

This caused artifacts when packet loss occured as the payload header
only existed in the first packet when it in fact could have been there
for each packet.